requirements.txt
pandas
openpyxl
pyarrow
faiss-cpu
openai
langchain
python-dotenv